move ConventionManageConnectedUser component to convention folder
